feat(account): migrate self-service to modules/account (MON-569)
Folds profile, session, api-keys, organization, team, onboarding,
agent-settings, contact-settings into a new @modules/account package.
Deletes auth/onboarding/agent-settings/contact-settings repositories
(findMemberByUserId inlined into core/authentication). Adds minioClient
to ORPCContext so avatar/logo upload procedures live in the module.
oRPC middleware (requireOrganizationTeam) replaces ad-hoc helpers.
All routers <300 LOC; extracts cnpj.ts and onboarding-seed.ts to
keep onboarding under the limit.
