+++ apps/web/src/integrations/orpc/router/index.ts
@@ -1,14 +1,15 @@
-import * as accountRouter from "./account";
+import * as accountRouter from "@modules/account/router/profile";
 import * as rubiRouter from "@modules/agents/router/chat";
 import * as threadsRouter from "@modules/agents/router/threads";
-import * as agentSettingsRouter from "./agent-settings";
+import * as agentSettingsRouter from "@modules/account/router/agent-settings";
 import * as analyticsRouter from "./analytics";
-import * as apiKeysRouter from "./api-keys";
+import * as apiKeysRouter from "@modules/account/router/api-keys";
 import * as bankAccountsRouter from "./bank-accounts";
 import * as billingRouter from "@modules/billing/router/billing";
 import * as categoriesRouter from "@modules/classification/router/categories";
 import * as categoriesBulkRouter from "@modules/classification/router/categories-bulk";
-import * as contactSettingsRouter from "./contact-settings";
+import * as cnpjRouter from "@modules/account/router/cnpj";
+import * as contactSettingsRouter from "@modules/account/router/contact-settings";
 import * as contactsRouter from "@modules/billing/router/contacts";
 import * as couponsRouter from "@modules/billing/router/coupons";
 import * as creditCardsRouter from "./credit-cards";
@@ -18,8 +19,8 @@ import * as financialSettingsRouter from "./financial-settings";
 import * as insightsRouter from "./insights";
 import * as inventoryRouter from "./inventory";
 import * as notificationsRouter from "./notifications";
-import * as onboardingRouter from "./onboarding";
-import * as organizationRouter from "./organization";
+import * as onboardingRouter from "@modules/account/router/onboarding";
+import * as organizationRouter from "@modules/account/router/organization";
 import * as benefitsRouter from "@modules/billing/router/benefits";
 import * as metersRouter from "@modules/billing/router/meters";
 import * as pricesRouter from "@modules/billing/router/prices";
@@ -27,9 +28,9 @@ import * as servicesRouter from "@modules/billing/router/services";
 import * as subscriptionItemsRouter from "@modules/billing/router/subscription-items";
 import * as subscriptionsRouter from "@modules/billing/router/subscriptions";
 import * as usageRouter from "@modules/billing/router/usage";
-import * as sessionRouter from "./session";
+import * as sessionRouter from "@modules/account/router/session";
 import * as tagsRouter from "@modules/classification/router/tags";
-import * as teamRouter from "./team";
+import * as teamRouter from "@modules/account/router/team";
 import * as transactionsRouter from "./transactions";
 
 export default {
@@ -42,6 +43,7 @@ export default {
    creditCards: creditCardsRouter,
    categories: categoriesRouter,
    categoriesBulk: categoriesBulkRouter,
+   cnpj: cnpjRouter,
    contactSettings: contactSettingsRouter,
    contacts: contactsRouter,
    coupons: couponsRouter,
