+++ apps/web/src/integrations/orpc/server.ts
@@ -15,9 +15,11 @@ import { AppError, WebAppError } from "@core/logging/errors";
 import type { Redis } from "@core/redis/connection";
 import type { DBOSClient } from "@dbos-inc/dbos-sdk";
 import { sanitizeData } from "@core/utils/sanitization";
+import type { MinioClient } from "@core/files/client";
 import {
    auth,
    db,
+   minioClient,
    posthog,
    posthogPrompts,
    redis,
@@ -37,6 +39,7 @@ export interface ORPCContextWithAuth extends ORPCContext {
    posthogPrompts: Prompts;
    redis: Redis;
    workflowClient: DBOSClient;
+   minioClient: MinioClient;
 }
 
 export interface ORPCContextAuthenticated extends ORPCContextWithAuth {
@@ -67,6 +70,7 @@ const withDeps = base.use(async ({ context, next }) => {
          posthogPrompts,
          redis,
          workflowClient: await workflowClient,
+         minioClient,
       },
    });
 });
