+++ modules/account/package.json
@@ -1,0 +1,40 @@
+{
+   "name": "@modules/account",
+   "version": "0.1.0",
+   "private": true,
+   "license": "Apache-2.0",
+   "type": "module",
+   "exports": {
+      "./router/*": "./src/router/*.ts",
+      "./*": "./src/*.ts"
+   },
+   "scripts": {
+      "check": "oxlint ./src",
+      "format": "oxfmt --write ./src",
+      "format:check": "oxfmt --check ./src",
+      "test": "vitest run --passWithNoTests",
+      "typecheck": "tsgo"
+   },
+   "dependencies": {
+      "@core/authentication": "workspace:*",
+      "@core/database": "workspace:*",
+      "@core/files": "workspace:*",
+      "@core/logging": "workspace:*",
+      "@core/orpc": "workspace:*",
+      "@core/utils": "workspace:*",
+      "@modules/classification": "workspace:*",
+      "@orpc/server": "catalog:orpc",
+      "@packages/analytics": "workspace:*",
+      "dayjs": "catalog:ui",
+      "drizzle-orm": "catalog:database",
+      "drizzle-zod": "catalog:database",
+      "neverthrow": "catalog:validation",
+      "posthog-node": "catalog:analytics-client",
+      "zod": "catalog:validation"
+   },
+   "devDependencies": {
+      "@tooling/typescript": "workspace:*",
+      "typescript": "catalog:development",
+      "vitest": "catalog:testing"
+   }
+}
